alteração na variavel peso

+++ PastaFontes/Produto.java.java
@@ -2,27 +2,27 @@
 //Inclui e consulta dados de produto
 public class Produto {
 
-	private double peso;
+	private int peso;
 	private double altura;
 	
 	public Produto() {
 		
 	}
 	
-	public Produto(double peso, double altura) {
+	public Produto(int peso, double altura) {
 		this.peso=peso;
 		this.altura=altura;
 	}
 	
-	public double getPeso() {
+	public int getPeso() {
 		return peso;
 	}
 	
-	public void setPeso(double peso) {
+	public void setPeso(int peso) {
 		this.peso = peso;
 	}
 	
-	public double getAltura() {
+	public int getAltura() {
 		return altura;
 	}
